🎁 Apply custom styling for advanced search form
What started as an investigation into the advanced search form turned
into fixing css.  The advanced search is working as expected but the
form had some weird quirks that needed to be addressed.  This commit
will apply custom styling for the advanced search form so the form text
is easier to see.  Also the drop down menus will be more aligned.  As a
bonus, added a comment to the date_facet_range.scss file to align
documentation with the rest of the files.

Ref:
  - https://github.com/scientist-softserv/west-virginia-university/issues/99