🧹 Remove date from advanced search
There is a known issue with the date facet in advanced search and the
common approach is to remove it.  This commit will remove the date from
the advanced search form and also increase the character limit for the
form fields.  It was set to 4 but that was only enough for the year and
didn't work for a date search like DD/MM/YYYY.  Lastly, I moved the gem
call for `blacklight_range_limit` out of the development group so it can
be used in a deployed environment.  Other than those changes, the search
seems to be working as expected

Ref:
  - https://github.com/scientist-softserv/west-virginia-university/issues/100
  - https://github.com/LafayetteCollegeLibraries/spot/pull/143