🎁 Add xml exporter
This commit will add an xml exporter in addition to the csv exporter. A
bit of refactoring was needed to get the exporter to work.  We have to
use a different approach than using the CatalogController#index because
there's already a default feature when doing `/catalog.xml`.  Instead we
use an `#export` action for both exporters.  To DRY up the code we
introduce the `ExportResultsPresenter` to handle the shared logic.  Also
we add a little bit of CSS to make the button look like the others as
well as change the button to a dropdown menu.

Ref:
  - https://github.com/scientist-softserv/west-virginia-university/issues/94