🎁 Allows user to import identifiers with protocols and domains
In this commit, we add functionality to clean the identifier which should allow a user to import identifiers as uris in bulkrax. Previously the issue was that the Acda instance would fail to create because it assigns the identifier as the id. Now the protocol and domain bits will be stripped out, allowing the app to maintain its original functionality.

Issue:
- https://github.com/scientist-softserv/west-virginia-university/issues/140

Co-Authored-By: Kirk Wang <k3wang@gmail.com>