🎁 Modify the way date range limits work
This commit aligns the way that date range limits work with
Palni/Palci's implementation.  The original problem was that the date
parser wasn't particularly robust and would fail on dates that were not
of a certain format.  Here we take the more conformed dates from the
edtf_tesim field and strip the year out and sort it by that.  We now
limit the number of characters to only 4 so that only YYYY can be
searched.  The facet UI has been updated to make it more clear as well.
We redefine the date_ssi field to hold only the year and that is what we
search with.  This does not affect XML and CSV catalog result exports
because date_tesim is used there.

Ref:
  - https://github.com/scientist-softserv/west-virginia-university/issues/112