🐛 Fix bookmark export
This commit will address a bug in the bookmark export.  Previously, the
export button on the bookmarks page would export with effectively no
query params.  This is because the params did not have a `q` and
basically did a `q=""` search.  If you had 1 bookmark then it'll just
return the first result from an empty query search, which most likely
would not be the bookmarked record you wanted.

This also enables the user to pick specific results via the bookmark
feature and only export those results.