🎁 Generate thumbnails on update
This commit will allow the `GenerateThumbsJob` to be called on update if
the `preview` field has been changed.  We also handle errors better in
case we don't have a valid URL for thumbnails.  If we get a invalid URL
then the existing thumbnail will be removed to better indicate that
there was something wrong with the URL.

There is now a lot of duplication between the `GenerateImageThumbsJob`
and the `GeneratePdfThumbsJob`.  Perhaps a refactor in the future would
be in order.