🐛 Move database calls out of the views
This commit will fix a bug where the id was saved as the identifier.
Though not truly a bug but more of a misnomer, it however did affect URI
identifiers because identifiers were not the same as the id in this
case.

This commit will also add a new indexed field to check weather or not
the record has a thumbnail or image file.  This logic will allow us to
move away from the database call in the view and instead leverage the
Solr Document.