Update _document.html.erb
This commit will update the _document.erb any any affected parts of it
to use Blacklight 7's new view components paradigm.  One quirk that I
found was using the #link_to_facets on the catalog controller isn't
working as expected.  It gives an tag but it was a string and not an
html safe version where it actually behaves like an a tag.  I had to
implement a helper method #render_html_safe_url to get the links to
acutally render as links.

CSS has been updated to make the component look better and close as to
production as possible.  There are a few differences that we'll see if
we need to change in the future like how the fields are displayed.  It
used to be one line, field title and field value.  Now it's field title
on a line and then field value after a return carriage.