Fix date parsing
Previously the date_ssi did not allow for multiple dates which did not
work with a range of dates.  This commit will change it to date_ssim and
only use the year to better work with the range limit gem.  We are
assuming only EDTF dates are coming in and will not index if it is not
proper EDTF.