Add exceptions to validation service
We are adding 'undated' and 'unknown' to the EDTF and creator
validations respectively.  Since these are the only two exceptions,
we're fine with not changing the code too much to accommodate.  If in
the future, we are seeing more exceptions, then we would want to
refactor the validation service to handle exceptions better.

In addition, we are using the #is_active_url? method from
ApplicationHelper in the service, but we want to give the user the
option to either do a quick check (check if string starts with "http")
or a full check whether the URL resolves correctly or not.