Adjust Getty AAT search
This commit mainly will address an oddity where the sparql query for the
Getty AAT search will not work.  We're adding a backup web scrape (like
the the TGN) search to check for the term as well.  Also added more to
the LC Linked Data error to include which dataset was searched.